refactor(web): alert destination composers and event labelling live in web-shared/
composers.ts was identical in both frontends apart from the relative path to
the ntfy helper, which is now @web-shared/ntfy in both, so the move is a
straight relocation.

eventLabel and parseCsv are pure but were typed against i18next's TFunction,
which web-shared/ cannot resolve: it has no node_modules and each app resolves
packages from its own. The translator is therefore typed structurally, as the
key-plus-defaultValue lookup the label actually performs, and both apps' own
TFunction satisfies it. Leaving the .tsx files also drops the two
react-refresh/only-export-components disables they needed.

The severity palettes stay per app: the same four severities name different
colour tokens in the two design systems.
