+++ frontdesk/web/src/components/alerts/steps.tsx
@@ -1,3 +1,12 @@
+import {
+	APPRISE_SERVICES_URL,
+	type DestinationKind,
+	FIELDS,
+	type FieldDef,
+	parseDiscordWebhook,
+	parseUnifiedPushEndpoint,
+} from "@web-shared/alerts/composers";
+import { eventLabel } from "@web-shared/alerts/events";
 import { generateTopic } from "@web-shared/ntfy";
 import type { TFunction } from "i18next";
 import {
@@ -10,16 +19,8 @@ import {
 import type { AlertEventDef, AlertStatus } from "../../api/types";
 import { CopyRow } from "../CopyRow";
 import { type Action, isDuplicate, type WizardState } from "./AlertsWizard";
-import {
-	APPRISE_SERVICES_URL,
-	type DestinationKind,
-	FIELDS,
-	type FieldDef,
-	parseDiscordWebhook,
-	parseUnifiedPushEndpoint,
-} from "./composers";
 import { DestinationList } from "./DestinationList";
-import { eventLabel, SEVERITY_COLOR } from "./events";
+import { SEVERITY_COLOR } from "./events";
 
 // The seven step bodies of the alerts wizard: prove apprise-api answers, pick
 // what kind of destination this is, fill in the parts that kind needs, deliver
