+++ web/src/pages/Settings/AlertEventPicker.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import type { TFunction } from "i18next";
+import { eventLabel } from "@web-shared/alerts/events";
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { CheckSquare, Square } from "@/lib/icons";
@@ -16,16 +16,6 @@ const SEVERITY_DOT: Record<string, string> = {
 	warning: "var(--warning-text)",
 	error: "var(--error-text)",
 };
-
-// eventLabel names one catalog event. The dots in an event type are not legal
-// in a translation key, so they become underscores; an event the locale does
-// not name yet reads as its own type rather than as a missing key.
-// eslint-disable-next-line react-refresh/only-export-components
-export function eventLabel(t: TFunction, type: string): string {
-	return t(`settings.alerts.event.${type.replace(/\./g, "_")}`, {
-		defaultValue: type,
-	});
-}
 
 interface AlertEventPickerProps {
 	/** Current alert_events value (CSV); undefined when the key is unset (first run). */
