+++ web/src/pages/Settings/AlertSnippets.tsx
@@ -1,9 +1,9 @@
+import { APPRISE_SERVICES_URL } from "@web-shared/alerts/composers";
 import { useTranslation } from "react-i18next";
 import { ExternalLink } from "@/lib/icons";
 import type { LangIconKey } from "../../components/langIcons";
 import { ShikiCode } from "../../components/ShikiCode";
 import { TerminalPreview } from "../../components/TerminalPreview";
-import { APPRISE_SERVICES_URL } from "./alerts/composers";
 
 /**
  * Example Apprise URL shapes for the most popular services, shown as copyable
